py-be/pyproject.toml
[project]
name = "starkfinder"
version = "0.1.0"
description = "Onchain AI agent on Starknet for all sorts of people in the ecosystem"
authors = [
    {name = "Soham Das",email = "todrsoham@gmail.com"}
]
readme = "./../README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "prisma (>=0.15.0,<0.16.0)",
    "starknet-py (>=0.27.0,<0.28.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.3.0"
psycopg2-binary = "^2.9.10"
sqlalchemy-utils = "^0.41.2"

[tool.isort]
profile = "black"
line_length = 88
